Update from internal code

+++ db/migrate/20130105201537_create_sales_tax_rates.rb
@@ -3,7 +3,7 @@ class CreateSalesTaxRates < ActiveRecord::Migration
     create_table :sales_tax_rates do |t|
       t.integer :sales_tax_customer_class_id
       t.integer :sales_tax_product_class_id
-      t.float :rate
+      t.decimal :rate
 
       t.timestamps
     end
